+++ packages/server/src/aggregates/session/domain/session.abstract.service.ts
@@ -1,1 +1,7 @@
-export abstract class SessionAbstractService {}
+import { Request } from 'express';
+import { SessionEntity } from '@rateme/core/domain/entities/session.entity';
+
+export abstract class SessionAbstractService {
+  abstract getSession(request: Request): Promise<SessionEntity>;
+}
+
